feat: add travis-ci fix: resolve bugs from the Seedling and SeedlingSummary test specs
