+++ .travis.yml
@@ -1,0 +1,13 @@
+language: node_js
+node_js:
+   - node # uses most recent stable node version
+cache:
+  directories:
+    - node_modules
+install:
+  - npm install
+notifications:
+  email:
+    on_success: change # default: change (only when going from broken to fixed)
+    on_failure: always # default: always (which is annoying, as it should be)
+
