--- src/App.spec.js
@@ -1,27 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './components/index';
-import { db, firebaseAuth } from './config/constants';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-
-const muiTheme = getMuiTheme(
-  {
-    "palette": {
-      "primary1Color": "#4caf50",
-      "primary2Color": "#689f38",
-      "accent1Color": "#b2ff59",
-      "pickerHeaderColor": "#009688"
-    },
-    "textField": {
-      "errorColor": "#ff5722"
-    }
-  });
-
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(
-    <MuiThemeProvider muiTheme={getMuiTheme(muiTheme)}>
-      <App />
-    </MuiThemeProvider>, div);
-});
